Renamed iterator and view files for iterating mapped objects
